feat!: rename to interceptor-manager
