+++ CHANGELOG.md
@@ -5,12 +5,12 @@
 
 ### Features
 
-* hooks migration from interceptor repo to the separate one ([#3](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor.hook-samples/issues/3)) ([2a2841a](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor.hook-samples/commit/2a2841afd2fa6f23fcd6ece66f0db1bed88eebab))
-* migration to interceptor v2.1.0 ([#12](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor.hook-samples/issues/12)) ([bf70a76](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor.hook-samples/commit/bf70a7607efb36ffa2eda8486f2c6d51de2d8ef7))
+* hooks migration from interceptor repo to the separate one ([#3](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor-manager.hook-samples/issues/3)) ([2a2841a](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor-manager.hook-samples/commit/2a2841afd2fa6f23fcd6ece66f0db1bed88eebab))
+* migration to interceptor v2.1.0 ([#12](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor-manager.hook-samples/issues/12)) ([bf70a76](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor-manager.hook-samples/commit/bf70a7607efb36ffa2eda8486f2c6d51de2d8ef7))
 
 
 ### Bug Fixes
 
-* **deps:** update dependency ch.sbb.polarion.extensions:ch.sbb.polarion.extension.interceptor to v2.2.0 ([c152a5b](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor.hook-samples/commit/c152a5b6d73b5bf407ba2c7245e0c1695c1b7453))
-* display hook version in Admin pane ([#16](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor.hook-samples/issues/16)) ([e6efcde](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor.hook-samples/commit/e6efcdeed2a1cb093dc1ef74df6e96b6b5c17679))
-* fix cast exception ([#7](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor.hook-samples/issues/7)) ([06db530](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor.hook-samples/commit/06db53031359981986f96a324a8fc122fea4c7e9))
+* **deps:** update dependency ch.sbb.polarion.extensions:ch.sbb.polarion.extension.interceptor to v2.2.0 ([c152a5b](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor-manager.hook-samples/commit/c152a5b6d73b5bf407ba2c7245e0c1695c1b7453))
+* display hook version in Admin pane ([#16](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor-manager.hook-samples/issues/16)) ([e6efcde](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor-manager.hook-samples/commit/e6efcdeed2a1cb093dc1ef74df6e96b6b5c17679))
+* fix cast exception ([#7](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor-manager.hook-samples/issues/7)) ([06db530](https://github.com/SchweizerischeBundesbahnen/ch.sbb.polarion.extension.interceptor-manager.hook-samples/commit/06db53031359981986f96a324a8fc122fea4c7e9))
