+++ hook-samples/only-assignee-can-delete/src/main/java/ch/sbb/polarion/extension/interceptor_manager/hook_samples/only_assignee_can_delete/OnlyAssigneeCanDeleteHook.java
@@ -1,8 +1,8 @@
-package ch.sbb.polarion.extension.interceptor.hook_samples.only_assignee_can_delete;
+package ch.sbb.polarion.extension.interceptor_manager.hook_samples.only_assignee_can_delete;
 
-import ch.sbb.polarion.extension.interceptor.model.ActionHook;
-import ch.sbb.polarion.extension.interceptor.model.HookExecutor;
-import ch.sbb.polarion.extension.interceptor.util.PropertiesUtils;
+import ch.sbb.polarion.extension.interceptor_manager.model.ActionHook;
+import ch.sbb.polarion.extension.interceptor_manager.model.HookExecutor;
+import ch.sbb.polarion.extension.interceptor_manager.util.PropertiesUtils;
 import com.polarion.alm.projects.model.IUser;
 import com.polarion.alm.tracker.ITrackerService;
 import com.polarion.alm.tracker.model.IWorkItem;
