+++ hook-samples/single-assignee/src/main/resources/META-INF/MANIFEST.MF
@@ -1,2 +1,2 @@
 Name: Single assignee hook for Polarion ALM Interceptor plugin
-Main-Class: ch.sbb.polarion.extension.interceptor.hook_samples.single_assignee.SingleAssigneeHook
+Main-Class: ch.sbb.polarion.extension.interceptor_manager.hook_samples.single_assignee.SingleAssigneeHook
