+++ hook-samples/testrun/src/main/resources/META-INF/MANIFEST.MF
@@ -1,2 +1,2 @@
 Name: TestRun validation hook for Polarion ALM Interceptor plugin
-Main-Class: ch.sbb.polarion.extension.interceptor.hook_samples.testrun.TestRunHook
+Main-Class: ch.sbb.polarion.extension.interceptor_manager.hook_samples.testrun.TestRunHook
