+++ hook-samples/title-length-check/src/main/java/ch/sbb/polarion/extension/interceptor_manager/hook_samples/title_length_check/TitleLengthHook.java
@@ -1,8 +1,8 @@
-package ch.sbb.polarion.extension.interceptor.hook_samples.title_length_check;
+package ch.sbb.polarion.extension.interceptor_manager.hook_samples.title_length_check;
 
-import ch.sbb.polarion.extension.interceptor.model.ActionHook;
-import ch.sbb.polarion.extension.interceptor.model.HookExecutor;
-import ch.sbb.polarion.extension.interceptor.util.PropertiesUtils;
+import ch.sbb.polarion.extension.interceptor_manager.model.ActionHook;
+import ch.sbb.polarion.extension.interceptor_manager.model.HookExecutor;
+import ch.sbb.polarion.extension.interceptor_manager.util.PropertiesUtils;
 import com.polarion.alm.tracker.model.IModule;
 import com.polarion.alm.tracker.model.IWorkItem;
 import com.polarion.core.util.logging.Logger;
