+++ hook-samples/title-length-check/src/main/resources/META-INF/MANIFEST.MF
@@ -1,2 +1,2 @@
 Name: Title length hook for Polarion ALM Interceptor plugin
-Main-Class: ch.sbb.polarion.extension.interceptor.hook_samples.title_length_check.TitleLengthHook
+Main-Class: ch.sbb.polarion.extension.interceptor_manager.hook_samples.title_length_check.TitleLengthHook
